add get and set in busnnis class

+++ EntityLayer/clsEntityCircle.cs
@@ -5,6 +5,6 @@
         public int CircleID { get; set; } = -1;
         public string CircleName { get; set; } = string.Empty;
         public int TeacherID { get; set; } = -1;
-        public int MaxCapacity { get; set; } = 0;
+        public sbyte MaxCapacity { get; set; } = 0;
     }
 }
